move `InlineModePanel` into its own file
