+++ client/src/inlineMode.tsx
@@ -1,21 +1,11 @@
-import React, { useEffect, useMemo, useRef, useState } from "react";
-import {
-  INLINE_TOGGLE_WIDTH,
-  InlinePinboardTogglePortal,
-} from "./inlinePinboardToggle";
+import React, { useEffect, useMemo, useState } from "react";
+import { InlinePinboardTogglePortal } from "./inlinePinboardToggle";
 import { useLazyQuery } from "@apollo/client";
 import { gqlGetItemCounts } from "../gql";
 import { PinboardIdWithItemCounts } from "../../shared/graphql/graphql";
-import { css, Global } from "@emotion/react";
-import { neutral } from "@guardian/source-foundations";
-import { boxShadow, highlightItemsKeyFramesCSS } from "./styling";
-import { pinboard } from "../colours";
-import { Pinboard } from "./pinboard";
 
 export const WORKFLOW_TITLE_QUERY_SELECTOR =
   ".content-list-item__field--pinboard";
-
-export const INLINE_PANEL_WIDTH = 260;
 
 interface InlineModeProps {
   workflowTitleElements: HTMLElement[];
@@ -83,47 +73,5 @@ export const InlineMode = ({ workflowTitleElements }: InlineModeProps) => {
         )
       )}
     </React.Fragment>
-  );
-};
-
-interface InlineModePanelProps {
-  pinboardId: string;
-}
-
-export const InlineModePanel = ({ pinboardId }: InlineModePanelProps) => {
-  const panelRef = useRef(null);
-
-  return (
-    <div
-      onClick={(e) => e.stopPropagation()}
-      ref={panelRef}
-      css={css`
-        position: fixed;
-        display: flex;
-        flex-direction: column;
-        z-index: 3;
-        top: 100px;
-        bottom: 5px;
-        left: ${(document
-          .querySelector(".content-list-head__heading--pinboard")
-          ?.getBoundingClientRect()?.left || 0) +
-        INLINE_TOGGLE_WIDTH +
-        25}px;
-        background: ${neutral[93]};
-        box-shadow: ${boxShadow};
-        width: ${INLINE_PANEL_WIDTH}px;
-        border: 3px solid ${pinboard["500"]};
-        border-radius: 5px;
-      `}
-    >
-      <Global styles={highlightItemsKeyFramesCSS} />
-
-      <Pinboard
-        pinboardId={pinboardId}
-        isSelected
-        isExpanded
-        panelElement={panelRef.current}
-      />
-    </div>
   );
 };
